refactor: extract routes and controller logic
